+++ src/sys/webgl/context.rs
@@ -151,6 +151,15 @@ impl BitOrAssign for ClearMask {
 }
 
 
+#[repr(u32)]
+#[non_exhaustive]
+#[derive(Clone, Copy, Debug)]
+pub enum FramebufferAttachment {
+  Color = WebGl2RenderingContext::COLOR_ATTACHMENT0,
+  Depth = WebGl2RenderingContext::DEPTH_ATTACHMENT,
+}
+
+
 #[derive(Debug, Eq, PartialEq)]
 pub struct FramebufferStatus(u32);
 
@@ -317,6 +326,7 @@ impl Gl for Context {
   type ClearMask = ClearMask;
   type CullFace = CullFace;
   type Factor = Factor;
+  type FramebufferAttachment = FramebufferAttachment;
   type FramebufferStatus = FramebufferStatus;
   type FrontFace = FrontFace;
   type Func = Func;
@@ -462,11 +472,16 @@ impl Gl for Context {
   }
 
   #[inline]
-  fn set_framebuffer_depth_texture(&self, texture_target: TextureTarget, texture: &Texture) {
+  fn set_framebuffer_texture(
+    &self,
+    attachment: FramebufferAttachment,
+    texture_target: TextureTarget,
+    texture: &Texture,
+  ) {
     let mipmap_level = 0;
     let () = self.0.framebuffer_texture_2d(
       WebGl2RenderingContext::FRAMEBUFFER,
-      WebGl2RenderingContext::DEPTH_ATTACHMENT,
+      attachment as _,
       texture_target as _,
       Some(texture),
       mipmap_level,